detach volume without events prior to removal

+++ app/Jobs/Kingpin/Volume/Undeploy.php
@@ -28,6 +28,10 @@ class Undeploy extends Job
             return;
         }
 
+        if ($this->model->instances()->count() !== 0) {
+            throw new \Exception('Volume ' . $this->model->id . ' had instances when trying to delete');
+        }
+
         try {
             $this->model->availabilityZone->kingpinService()->delete(
                 '/api/v1/vpc/' . $this->model->vpc->id . '/volume/' . $this->model->vmware_uuid
